Enforce minimum number of sampled points
Previously if you queried for correlations of a monthly data set, it
would frequently find very short overlaps at the ends with yearly data
sets with only one point in the overlap period and the rest
interpolated, which looked very boring. This enforces the minimum
number of sampled points on the other data set as well as the query.